Threesome clause for Yaoi and Yuri in d/d/1
In most cases of visual novels that I have applied the Yuri category to,
I did so because 2 (or more) members of the female sex were involved in
a sexual act together, most cases pleasing a member of the male sex
together.

I believe this is the correct usage of this category, and does not prove
that the members of the female sex are (romantically) in love with each
other, which is the purpose of the Shoujo Ai category.

By extension, this same reasoning applys to Yaoi and Shounen Ai.