Replacing all occurences of 'discussion tags' with 'discussion boards'
Which is a more accurate description, and doesn't confuse with the
tagging system.

Note than even all internal uses of the word 'tag' have been replaced,
as I'm not a huge fan of different terminology in the code and UI. This
update might break some things related to the discussion board.