Rewrote Multi::Sitemap and included tag pages
This should've taken me less than an hour, but unfortunately it took a
bit longer because I found a rather hard-to-fix bug in PoCo::Pg.
Be sure to update to the latest POE::Component::Pg.

Also, this change removes the DateTime dependency. Its functionality
could be replaced with the strftime() found in POSIX (which every Perl
has)