L10N: Converted homepage and history pages
There are a few things left hardcoded, though: The 'expand' link is also
in Javascript, that will probably need some mechanism for TL as well.

The format of the items on the homepage might need to change depending
on the language, I can imagine that not every language uses the 'title
by name' construct. Especially the age() function needs a localized
version.