L10N: Converted Handler::Producers
Hmmm, I think I'll need to update some of the utility functions to take
more functionality from the Handler pages in regard to getting text
strings. Using mt() on everything that is essentially already passed to
other functions with a unique string for identification is a bit double.

So far I've updated htmlRevision to do this. (The name argument is
still present but currently ignored, to preserve backwards compatility,
the rest of the code will be updated soon.) Other utility functions like
htmlForm should probably do something like that as well.