Wrote LangFile module to handle lang.txt files
Figured I'd need to do some automated processing on that file in the
future, and considering the format was made to be easy to handle for
humans (and not computers), this module will come in handy.

I wrote that small POD for brainstorming the API, before typing even a
single line of code.