Wrote a quick and dirty translation editor plugin
See Plugin::TransAdmin for details.

I wrote it as a plugin so that it can make use of the framework and
functions of the main VNDB website code, but that it doesn't have to
confirm to the rest of the code or have to be loaded into memory when
unused.