Multi::Core: Improved logging of warnings and errors
Rather than forwarding STDERR and STDOUT to /dev/null, it is now written
to the log file. This also means that exceptions are now finally logged
somewhere, making debugging a lot easier.

I've also looked at how we should handle exceptions, but the default
action of POE is to let the process die() as any normal script (assuming
a session doesn't sig_handled() the DIE signal), and I couldn't think of
anything better than this default - so let's code with this in mind:
die() when something goes horribly wrong, otherwise just warn() and
handle the error.