Multi::API: Rewrote filtertosql()
The filters are now defined in the arguments rather than the function,
the function simply provides the necessary means to define the filters
in a concise way.

This also means there's a lot less error handling code necessary, and
therefore merged the 'filter*' error messages into one 'filter' error.

This code is far more readable and maintainable than the previous
implementation.