Auth: Combined dbSessionCheck into dbUserGet
This one query is a bit faster than the two queries executed seperately, and
with a query that is executed on each pageview it does matter.

Ideally, the dbUserMessageCount() is cached and fetched with the same query,
this should save another 1-2ms. But this is probably not worth the extra code
it would require.