Multi::Core: Call $poe_kernel->has_forked() after forking
Normally this shouldn't be necessary as the forking is done before any
POE::Kernel functions has been called. It seems, however, that the
POE::Kernel is already started in the import() function, meaning that
it's already running anyway.

Not sure whether the import() has changed to start the kernel or that
something else changed that I'm only seeing this now, ow well.