SQL: Removed changes.causedby
This column was used to differentiate between automated edits and user
edits, but that later changed to checking for changes.requester = 1.

The column has since never really been used, and due to a bug introduced
in VNDB 2.0, it has never been updated, either. Meaning it's not even
accurate for any database changes made after december 2008...