Added simple SQL manipulation and upgrade functionality to the Makefile
This makes it possible to do (mostly) automated updates, and makes sure
I don't forget to update something, as usual.

Of course, this update functionality is not fool-proof, and in no way
subsistutes for manually checking what an update requires. It simply
does the things that a basic script can do without relying on a specific
configuration.