Versioned the deleting and locking of database entries
This is implemented by adding ihid (item hidden) and ilock (item
locked) columns to the changes table,
The (vn|release|producer).(hidden|locked) columns now work as a
cache, refering to the changes.(ihid|ilock) columns with
  changes.id = (vn|release|producer).latest.

The cached columns are updated automatically each time a new revision is
inserted.

This is a pretty large change, bugs are quite likely.