VNDB::Func: Another gtintype() related bugfix
We have to assume that the input doesn't always have enough zeros,
to fill the 12 characters, otherwise UPC codes fail when they are
fetched from the database (which is stored as a bigint, no zero padding
in the output). Instead, we'll just add the zeros ourselves up to 12 for
normalization purposes.