Handler::(Misc|VNPage): Added HTML width/height to screenshot thumbnails
And made sure the dimensions are truncated in VNDBUtil::imgsize().

Setting the width/height attributes makes sure that the browser can
reserve space for the image when it hasn't been loaded yet, which
prevents the overall page layout from changing while the images are
loading. (which is annoying if your connection isn't all that fast)