VNDBUtil: Partly rewrote bb2html() to be faster and better
It does not use split() anymore, the input string is parsed in a single
pass using a global regex. It's now a lot faster on larger input. The
page generation time of /t937 went back from ~350ms to ~55ms (on the
beta, the production server is slower).

Also made several tiny improvements while I was at it:
- multiple successive newlines aren't removed within [code]
- truncating a message with $maxlength also removes trailing
  spaces and interpunction
- multiple successive spaces are removed outside of [code] (and thus
  don't count towards the length of the message)

The function should be mostly equivalent in regards to all other things,
ignoring a few minor border cases that weren't documented in the first
place. The URL regex (and the idea of a global regex) came from bpaste:
  http://g.blicky.net/bpaste.git/commit/?id=ac7b16d0ec0f195d00a0a79698f67c3010e8cf7d