Re-structured password recovery feature
Rather than setting an automatically password, reset the password and
send an email with a secure token instead. The password can then be set
again using this token.

This doesn't really have an advantage at this point, just makes the
interface and code more consistent when I update the registration code
to do something similar.