Multi::API: Tighten input validation for integers
This fixes a vulnerability where one could remotely crash Multi by
sending a large integer in a filter. PostgreSQL will throw an error if
an integer is out of range for the selected data type, and Multi shuts
itself down on an SQL error.

(I've known about this vulnerability for quite a while, but it's good to
see how nobody has exploited it yet.)