norm_ip: Use /23 subnet for IPv4
Apparently there are networks where a single visitor is assigned random
addresses in the /23 range for each request. This should fix the
login/registration form on such networks, and makes the login throttling
more robust (and easier to trigger for innocent people, but judging from
monitoring the throttle table, failed logins arent that common).

I wonder if /23 is enough, but we'll see.