Hash session tokens with SHA-1 when storing in DB
This ensures that, if an attacker evers gets read access to the
database, he will not be able to compromise any accounts. SHA-1 suffices
here, because the data being hashed is a random 20 byte string. The
search space is so damn large that you can't sanely brute force it, nor
are rainbow tables any use at that scale.

They're not salted. The password reset tokens are also hashed in the
database and do include salt, but I've no idea why we did that.