Multi::API: Support IPv6
This was more trivial than I had expected. I already took ipv6 into
account when rewriting the API for AnyEvent (including the use of
norm_ip()), so that part was fine. The only part I had to fix was the
listening socket, and I had to ensure that the $c->{ip} was correct. The
first was easy, and the latter was properly handled by AnyEvent
automatically. Looks like AnyEvent automatically 'unpacks' IPv4-mapped
IPv6 addresses, so I didn't have to deal with that myself.