Handler::VNPage: Rewrite release comparison page
The previous code was using experimental perl features (switch /
smartmatch) that weren't really needed, and the information about
individual columns was spread around in multiple functions. This rewrite
makes the code consistent with the rest of VNDB, and has *all* of the
column-specific information in one data structure.

I did not replicate the similar-cell-merging feature, partly because the
code for it is definitely not trivial and partly because it doesn't make
the table look any less cluttered. In fact, I feel that it only makes
the table harder to interpret because it looks messy. This is a matter
opinion, of course, so I might reimplement the feature if people who
actually use this comparison page want to.

Overall, I'm still undecided on whether this comparison page should
exist at all in its current form - it's not very user-friendly and often
looks cluttered. I'm keeping it because it does have some use-cases
where it avoids opening every release page to do a manual comparison,
but I'd love a more friendly-looking alternative.