Support zopfli/zopflipng for all static asset generators
Compresses a little better. I reduced the number of iterations required
to find the optimal image size in spritegen.pl, but generating the
icons.png is *incredibly slow* when combining zopflipng with the 'slow'
option. It's possible to parallelize the calculation and use multiple
cores to speed it up, but that seems overkill.

Some icons.png compression stats:

  METHOD            SIZE       RUNTIME
  default           18103      <1sec
  slow              17941      few secs
  pngcrush          15385      <1sec
  pngcrush+slow     15148      few mins
  zopflipng         14986      few secs
  zopflipng+slow    14898      ~1 hour